+++ source/standalone_examples/api/isaacsim.robot.manipulators/cobotta_900/gripper_control.py
@@ -30,7 +30,7 @@ my_world = World(stage_units_in_meters=1.0)
 assets_root_path = get_assets_root_path()
 if assets_root_path is None:
     raise Exception("Could not find Isaac Sim assets folder")
-asset_path = assets_root_path + "/Isaac/Robots/Denso/cobotta_pro_900.usd"
+asset_path = assets_root_path + "/Isaac/Robots/Denso/CobottaPro900/cobotta_pro_900.usd"
 add_reference_to_stage(usd_path=asset_path, prim_path="/World/cobotta")
 # define the gripper
 gripper = ParallelGripper(