+++ source/extensions/isaacsim.core.experimental.prims/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.1.1] - 2025-04-21
+### Changed
+- Update Isaac Sim robot asset path
+- Update docstrings for the Franka Panda robot
+
 ## [0.1.0] - 2025-03-21
 ### Added
 - Initial release